+++ src/premade_pools.rs
@@ -3,13 +3,16 @@
 //! These can be annoying due to orphan rules that prevent you from implementing your own methods,
 //! so feel free to copy-paste them (without attribution) into your own source to make new variants.
 
-use crate::pool::{MaxPoolLessThanZero, Pool};
+use crate::pool::{MaxPoolLessThanMin, Pool};
 use bevy::prelude::{Component, Resource};
-use core::ops::{Div, Mul};
+use core::fmt::{Display, Formatter};
+use core::ops::{Add, AddAssign, Div, Mul, Sub, SubAssign};
 use derive_more::{Add, AddAssign, Sub, SubAssign};
 
 /// A premade resource pool for life (aka health, hit points or HP).
 pub mod life {
+    use crate::pool::RegeneratingPool;
+
     use super::*;
 
     /// The amount of life available to a unit.
@@ -24,6 +27,25 @@ pub mod life {
         max: Life,
         /// The amount of life regenerated per second.
         pub regen_per_second: Life,
+    }
+
+    impl LifePool {
+        /// Creates a new [`LifePool`] with the supplied settings.
+        ///
+        /// # Panics
+        /// Panics if `current` is greater than `max`.
+        /// Panics if `current` or max is negative.
+
+        pub fn new(current: Life, max: Life, regen_per_second: Life) -> Self {
+            assert!(current <= max);
+            assert!(current >= LifePool::MIN);
+            assert!(max >= LifePool::MIN);
+            Self {
+                current,
+                max,
+                regen_per_second,
+            }
+        }
     }
 
     /// A quantity of life, used to modify a [`LifePool`].
@@ -58,21 +80,17 @@ pub mod life {
         }
     }
 
+    impl Div<Life> for Life {
+        type Output = f32;
+
+        fn div(self, rhs: Life) -> f32 {
+            self.0 / rhs.0
+        }
+    }
+
     impl Pool for LifePool {
         type Quantity = Life;
-        const ZERO: Life = Life(0.);
-
-        fn new(
-            current: Self::Quantity,
-            max: Self::Quantity,
-            regen_per_second: Self::Quantity,
-        ) -> Self {
-            LifePool {
-                current,
-                max,
-                regen_per_second,
-            }
-        }
+        const MIN: Life = Life(0.);
 
         fn current(&self) -> Self::Quantity {
             self.current
@@ -88,16 +106,18 @@ pub mod life {
             self.max
         }
 
-        fn set_max(&mut self, new_max: Self::Quantity) -> Result<(), MaxPoolLessThanZero> {
-            if new_max < Self::ZERO {
-                Err(MaxPoolLessThanZero)
+        fn set_max(&mut self, new_max: Self::Quantity) -> Result<(), MaxPoolLessThanMin> {
+            if new_max < Self::MIN {
+                Err(MaxPoolLessThanMin)
             } else {
                 self.max = new_max;
                 self.set_current(self.current);
                 Ok(())
             }
         }
+    }
 
+    impl RegeneratingPool for LifePool {
         fn regen_per_second(&self) -> Self::Quantity {
             self.regen_per_second
         }
@@ -106,10 +126,54 @@ pub mod life {
             self.regen_per_second = new_regen_per_second;
         }
     }
+
+    impl Add<Life> for LifePool {
+        type Output = Self;
+
+        fn add(mut self, rhs: Life) -> Self::Output {
+            self.set_current(self.current + rhs);
+            self
+        }
+    }
+
+    impl Sub<Life> for LifePool {
+        type Output = Self;
+
+        fn sub(mut self, rhs: Life) -> Self::Output {
+            self.set_current(self.current - rhs);
+            self
+        }
+    }
+
+    impl AddAssign<Life> for LifePool {
+        fn add_assign(&mut self, rhs: Life) {
+            self.set_current(self.current + rhs);
+        }
+    }
+
+    impl SubAssign<Life> for LifePool {
+        fn sub_assign(&mut self, rhs: Life) {
+            self.set_current(self.current - rhs);
+        }
+    }
+
+    impl Display for Life {
+        fn fmt(&self, f: &mut Formatter<'_>) -> core::fmt::Result {
+            write!(f, "{}", self.0)
+        }
+    }
+
+    impl Display for LifePool {
+        fn fmt(&self, f: &mut Formatter<'_>) -> core::fmt::Result {
+            write!(f, "{}/{}", self.current, self.max)
+        }
+    }
 }
 
 /// A premade resource pool for mana (aka MP).
 pub mod mana {
+    use crate::pool::RegeneratingPool;
+
     use super::*;
 
     /// The amount of mana available to a unit.
@@ -124,6 +188,24 @@ pub mod mana {
         max: Mana,
         /// The amount of mana regenerated per second.
         pub regen_per_second: Mana,
+    }
+
+    impl ManaPool {
+        /// Creates a new [`ManaPool`] with the supplied settings.
+        ///
+        /// # Panics
+        /// Panics if `current` is greater than `max`.
+        /// Panics if `current` or `max` is negative.
+        pub fn new(current: Mana, max: Mana, regen_per_second: Mana) -> Self {
+            assert!(current <= max);
+            assert!(current >= ManaPool::MIN);
+            assert!(max >= ManaPool::MIN);
+            Self {
+                current,
+                max,
+                regen_per_second,
+            }
+        }
     }
 
     /// A quantity of mana, used to modify a [`ManaPool`].
@@ -158,21 +240,17 @@ pub mod mana {
         }
     }
 
+    impl Div<Mana> for Mana {
+        type Output = f32;
+
+        fn div(self, rhs: Mana) -> f32 {
+            self.0 / rhs.0
+        }
+    }
+
     impl Pool for ManaPool {
         type Quantity = Mana;
-        const ZERO: Mana = Mana(0.);
-
-        fn new(
-            current: Self::Quantity,
-            max: Self::Quantity,
-            regen_per_second: Self::Quantity,
-        ) -> Self {
-            ManaPool {
-                current,
-                max,
-                regen_per_second,
-            }
-        }
+        const MIN: Mana = Mana(0.);
 
         fn current(&self) -> Self::Quantity {
             self.current
@@ -188,22 +266,66 @@ pub mod mana {
             self.max
         }
 
-        fn set_max(&mut self, new_max: Self::Quantity) -> Result<(), MaxPoolLessThanZero> {
-            if new_max < Self::ZERO {
-                Err(MaxPoolLessThanZero)
+        fn set_max(&mut self, new_max: Self::Quantity) -> Result<(), MaxPoolLessThanMin> {
+            if new_max < Self::MIN {
+                Err(MaxPoolLessThanMin)
             } else {
                 self.max = new_max;
                 self.set_current(self.current);
                 Ok(())
             }
         }
+    }
 
+    impl RegeneratingPool for ManaPool {
         fn regen_per_second(&self) -> Self::Quantity {
             self.regen_per_second
         }
 
         fn set_regen_per_second(&mut self, new_regen_per_second: Self::Quantity) {
             self.regen_per_second = new_regen_per_second;
+        }
+    }
+
+    impl Add<Mana> for ManaPool {
+        type Output = Self;
+
+        fn add(mut self, rhs: Mana) -> Self::Output {
+            self.set_current(self.current + rhs);
+            self
+        }
+    }
+
+    impl Sub<Mana> for ManaPool {
+        type Output = Self;
+
+        fn sub(mut self, rhs: Mana) -> Self::Output {
+            self.set_current(self.current - rhs);
+            self
+        }
+    }
+
+    impl AddAssign<Mana> for ManaPool {
+        fn add_assign(&mut self, rhs: Mana) {
+            self.set_current(self.current + rhs);
+        }
+    }
+
+    impl SubAssign<Mana> for ManaPool {
+        fn sub_assign(&mut self, rhs: Mana) {
+            self.set_current(self.current - rhs);
+        }
+    }
+
+    impl Display for Mana {
+        fn fmt(&self, f: &mut Formatter<'_>) -> core::fmt::Result {
+            write!(f, "{}", self.0)
+        }
+    }
+
+    impl Display for ManaPool {
+        fn fmt(&self, f: &mut Formatter<'_>) -> core::fmt::Result {
+            write!(f, "{}/{}", self.current, self.max)
         }
     }
 }